requirements.txt
graphiti-core
python-dotenv
neo4j
openai
anthropic
google-generativeai 